wrapped up optimizations

+++ libs/sosix/stdlib/posix/posixmemalign.c
@@ -29,7 +29,7 @@ int posix_memalign(void **memptr, size_t alignment, size_t size) {
 	int* memaddress = (int*)kalloc(1);
 
 	//lets turn memaddress to a value
-	uint64_t addr = (uint64_t)memaddress;
+	intptr_t addr = (intptr_t)memaddress;
 
 	mem_aligned_ctrl_block block;
